requirements.txt
Flask>=3.0.0
openai>=1.40.0
python-dotenv>=1.0.1
# Optional but enabled: lightweight GUI and web control
PyAutoGUI>=0.9.54
Pillow>=10.0.0
python-xlib>=0.33
